makefile & library working

+++ cursus/libft/strlen.c
@@ -6,12 +6,12 @@
 /*   By: daviles- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/08 13:44:19 by daviles-          #+#    #+#             */
-/*   Updated: 2023/03/13 19:29:28 by daviles-         ###   ########.fr       */
+/*   Updated: 2023/03/22 19:20:06 by daviles-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include"libft.h"
 
-unsigned long	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	int	count;
